Finished copy v0.1
Many changes that were not uploaded, ended up changing gears and designing it around the WHOIS creation date. If you run this daily and only raise an alert on script output, this script will work well for something like Datto RMM to give you a heads up that someone is impersonating one of the domains you protect.
I may later add in some checks for dnstwist installed with pip, but it does the job for me, for now.