Update twistwrapper.py v0.2
Fixed several things.
Now we save all the json output for review, but also maintain a log file that can be used to track progress of the script and how long it takes to run. For Datto RMM purpose, keeping it to not displaying anything but matching whois records. With a seven day lookback and considering the script runs a long time on a long list, running once a week might be sufficient, enough time to hopefully alert the customer and work with the registrar to have it taken down.